Removed the duplicate of scl_loader.py, added get_Data_Type_Definition
